rename utility types file
