+++ software/tracksight/frontend/src/types/Utility.ts
